Revert "Merge"
This reverts commit 557bb291cabb7a8ee224103979350bf041ea6208, reversing
changes made to b7bfcf9875dfdafc4471dd685bdb40a5a133eb72.

+++ eFurnitureProject.Domain/Entities/User.cs
@@ -13,7 +13,7 @@ namespace eFurnitureProject.Domain.Entities
         public DateTime? DateOfBird { get; set; }
         public string? Gender { get; set; }
         public string? Address { get; set; }
-        public int? Wallet { get; set; }
+        public double? Wallet { get; set; }
         public bool Status { get; set; } = true;
         
         public virtual ICollection<AppointmentDetail>? AppointmentDetail { get; set; }
